world: Fix Gold incorrectly having size 8 instead of 4

+++ wow_message_parser/src/parser/types/sizes.rs
@@ -42,7 +42,7 @@ pub(crate) const UPDATE_MASK_MIN_SIZE: u8 = 1;
 pub(crate) const PACKED_GUID_MAX_SIZE: u8 = 9;
 pub(crate) const PACKED_GUID_MIN_SIZE: u8 = 2;
 pub(crate) const GUID_SIZE: u8 = core::mem::size_of::<u64>() as u8;
-pub(crate) const GOLD_SIZE: u8 = core::mem::size_of::<u64>() as u8;
+pub(crate) const GOLD_SIZE: u8 = core::mem::size_of::<u32>() as u8;
 pub(crate) const DATETIME_SIZE: u8 = 4;
 
 impl Sizes {
